Implement a Namingprovider to override default Naminglogic
