+++ XmlSchemaClassGenerator.Console/Program.cs
@@ -91,7 +91,7 @@ If no mapping is found for an XML namespace, a name is generated automatically (
             {
                 var xn = key.XmlSchemaNamespace;
                 var name = string.Join(".", xn.Split('/').Where(p => Regex.IsMatch(p, @"^[A-Za-z]+$") && p != "schema")
-                    .Select(n => Generator.ToTitleCase(n, NamingScheme.PascalCase)));
+                    .Select(n => n.ToTitleCase(NamingScheme.PascalCase)));
                 if (!string.IsNullOrEmpty(namespacePrefix)) { name = namespacePrefix + (string.IsNullOrEmpty(name) ? "" : ("." + name)); }
                 return name;
             });
